text optimization

+++ src/landmark.c
@@ -389,7 +389,7 @@ static const struct LandmarkList sLandmarkLists[] =
 
 static const struct Landmark *const *GetLandmarks(u8 mapSection, u8 id);
 
-const u8 *GetLandmarkName(u8 mapSection, u8 id, u8 count)
+const u8 *GetLandmarkName(u8 mapSection, u8 id, u32 count)
 {
     const struct Landmark *const *landmarks = GetLandmarks(mapSection, id);
